feat(order_management): add current order section, menu selection section, and error banner
- Implemented CurrentOrderSection widget for displaying and managing current orders.
- Created MenuSelectionSection widget for selecting menu items with search and category filtering.
- Added OrderPageErrorBanner widget to display error messages with retry option.
- Developed unit tests for MenuSelectionSection and CurrentOrderSection to ensure functionality and interaction.
