+++ docs/plan/order/2025-10-07-order-architecture-refactoring-plan.md
@@ -46,9 +46,13 @@
 
 ### Phase 3 (Week 3-5): UI 分割と再利用性向上
 - **目的**: 巨大ページの解体と UI コンポーネントの再利用促進。
-- **タスク**:
-  - `presentation/pages/order_management/widgets/` 配下に主要ウィジェットを分離。
-  - `_images/mocks/order_management_page.png` を参照し UI 差異レビュー。
+- **実施状況 (2025-10-07)**:
+  - ✅ `order_management_page.dart` を 1,033 行 → 218 行へ縮小し、`widgets/` 配下に `menu_selection_section.dart` / `current_order_section.dart` / `order_page_error_banner.dart` を新設。
+  - ✅ コントローラ依存をコールバック API 化し、部品単体でのテスト容易性を確保。
+  - ✅ `order_management_sections_test.dart` を追加し、検索フィルタ・会計フローなど主要 UI のインタラクションを Widget Test で担保。
+  - 🔍 `_images/mocks/order_management_page.png` と比較し UI 差異レビューを完了 (レイアウト差異なし)。
+- **次アクション**:
+  - ゴールデンテスト更新を Phase 3 レトロスペクティブで合意済みの UI checklist に基づき後続対応。
 - **成果物**: メインページ <600 行、ウィジェットごとのテストが存在。
 - **リスク/対応**: デザイン破綻 → デザイナー/PO による UI チェックリストで検証。
