+++ lib/features/order/presentation/pages/order_management_page.dart
@@ -1,27 +1,21 @@
 import "dart:async";
-import "dart:math" as math;
 
 import "package:flutter/material.dart";
 import "package:flutter_riverpod/flutter_riverpod.dart";
 import "package:go_router/go_router.dart";
 
 import "../../../../shared/components/buttons/icon_button.dart";
-import "../../../../shared/components/inputs/quantity_stepper.dart";
-import "../../../../shared/components/inputs/search_field.dart";
-import "../../../../shared/components/inputs/segmented_filter.dart";
 import "../../../../shared/components/layout/page_container.dart";
-import "../../../../shared/components/layout/section_card.dart";
 import "../../../../shared/foundations/tokens/color_tokens.dart";
-import "../../../../shared/foundations/tokens/radius_tokens.dart";
 import "../../../../shared/foundations/tokens/spacing_tokens.dart";
-import "../../../../shared/foundations/tokens/typography_tokens.dart";
 import "../../../../shared/mixins/route_aware_refresh_mixin.dart";
-import "../../../../shared/patterns/patterns.dart";
+import "../../../../shared/patterns/navigation/app_top_bar.dart";
 import "../../../settings/presentation/pages/settings_page.dart";
 import "../controllers/order_management_controller.dart";
 import "../controllers/order_management_state.dart";
-import "../performance/order_management_tracing.dart";
-import "../widgets/order_payment_method_selector.dart";
+import "../widgets/order_management/current_order_section.dart";
+import "../widgets/order_management/menu_selection_section.dart";
+import "../widgets/order_management/order_page_error_banner.dart";
 import "order_status_page.dart";
 
 /// 注文管理画面のメインページ。
@@ -188,7 +182,7 @@ class _OrderManagementPageState extends ConsumerState<OrderManagementPage>
             ),
             if (state.isLoading) const SizedBox(height: YataSpacingTokens.md),
             if (state.errorMessage != null) ...<Widget>[
-              _OrderPageErrorBanner(message: state.errorMessage!, onRetry: controller.refresh),
+              OrderPageErrorBanner(message: state.errorMessage!, onRetry: controller.refresh),
               const SizedBox(height: YataSpacingTokens.md),
             ],
             Expanded(
@@ -197,16 +191,27 @@ class _OrderManagementPageState extends ConsumerState<OrderManagementPage>
                 children: <Widget>[
                   Expanded(
                     flex: 2,
-                    child: _MenuSelectionSection(
+                    child: MenuSelectionSection(
                       state: state,
-                      controller: controller,
                       searchController: _searchController,
+                      onSearchQueryChanged: controller.updateSearchQuery,
+                      onSelectCategory: controller.selectCategory,
+                      onUpdateItemQuantity: controller.updateItemQuantity,
+                      onAddMenuItem: controller.addMenuItem,
                     ),
                   ),
                   const SizedBox(width: YataSpacingTokens.lg),
                   // 右ペインはカード内で内部スクロール + 下部固定バー
                   Expanded(
-                    child: _CurrentOrderSection(state: state, controller: controller),
+                    child: CurrentOrderSection(
+                      state: state,
+                      onUpdateItemQuantity: controller.updateItemQuantity,
+                      onRemoveItem: controller.removeItem,
+                      onPaymentMethodChanged: controller.updatePaymentMethod,
+                      onOrderNotesChanged: controller.updateOrderNotes,
+                      onClearCart: controller.clearCart,
+                      onCheckout: controller.checkout,
+                    ),
                   ),
                 ],
               ),
@@ -216,817 +221,4 @@ class _OrderManagementPageState extends ConsumerState<OrderManagementPage>
       ),
     );
   }
-}
-
-class _MenuSelectionSection extends StatefulWidget {
-  const _MenuSelectionSection({
-    required this.state,
-    required this.controller,
-    required this.searchController,
-  });
-
-  final OrderManagementState state;
-  final OrderManagementController controller;
-  final TextEditingController searchController;
-
-  @override
-  State<_MenuSelectionSection> createState() => _MenuSelectionSectionState();
-}
-
-class _MenuSelectionSectionState extends State<_MenuSelectionSection> {
-  final ScrollController _scrollController = ScrollController();
-
-  @override
-  void dispose() {
-    _scrollController.dispose();
-    // _notesController.dispose();
-    super.dispose();
-  }
-
-  @override
-  Widget build(BuildContext context) {
-    final OrderManagementState state = widget.state;
-    final OrderManagementController controller = widget.controller;
-    final TextEditingController searchController = widget.searchController;
-    final List<YataFilterSegment> segments = state.categories
-        .map((MenuCategoryViewData category) => YataFilterSegment(label: category.label))
-        .toList(growable: false);
-
-    return YataSectionCard(
-      title: "メニュー選択",
-      subtitle: "商品をタップして注文に追加",
-      expandChild: true,
-      child: Column(
-        crossAxisAlignment: CrossAxisAlignment.start,
-        children: <Widget>[
-          YataSearchField(
-            controller: searchController,
-            hintText: "メニュー検索...",
-            onChanged: controller.updateSearchQuery,
-          ),
-          const SizedBox(height: YataSpacingTokens.md),
-          YataSegmentedFilter(
-            segments: segments,
-            selectedIndex: state.selectedCategoryIndex,
-            onSegmentSelected: controller.selectCategory,
-          ),
-          const SizedBox(height: YataSpacingTokens.lg),
-          // * メニュー一覧エリア（内部スクロール）
-          Expanded(
-            child: LayoutBuilder(
-              builder: (BuildContext context, BoxConstraints constraints) {
-                int itemCount = 0;
-                int sampledCrossAxisCount = 1;
-                double sampledTileHeight = 0;
-
-                return OrderManagementTracer.traceSync<Widget>(
-                  "page.menuGrid.build",
-                  () {
-                    final List<MenuItemViewData> items = state.filteredMenuItems;
-                    itemCount = items.length;
-                    if (state.isLoading) {
-                      return const Center(child: CircularProgressIndicator());
-                    }
-                    if (items.isEmpty) {
-                      return Center(
-                        child: Padding(
-                          padding: const EdgeInsets.symmetric(vertical: YataSpacingTokens.lg),
-                          child: Text(
-                            "メニューが見つかりません",
-                            style: Theme.of(
-                              context,
-                            ).textTheme.bodyMedium?.copyWith(color: YataColorTokens.textSecondary),
-                          ),
-                        ),
-                      );
-                    }
-
-                    const double spacing = YataSpacingTokens.xl;
-                    const double scrollbarThickness = 8;
-                    const double scrollPad = scrollbarThickness + YataSpacingTokens.xs;
-                    final double availableWidth = (constraints.maxWidth - scrollPad).clamp(
-                      0,
-                      double.infinity,
-                    );
-
-                    final int crossAxisCount = availableWidth >= 900
-                        ? 3
-                        : availableWidth >= 600
-                        ? 2
-                        : 1;
-                    final double itemWidth = crossAxisCount <= 0
-                        ? availableWidth
-                        : (availableWidth - spacing * (crossAxisCount - 1)) / crossAxisCount;
-                    final double tileHeight = _estimateMenuTileHeight(itemWidth, context);
-                    final int safeCrossAxisCount = crossAxisCount <= 0 ? 1 : crossAxisCount;
-                    sampledCrossAxisCount = safeCrossAxisCount;
-                    sampledTileHeight = tileHeight;
-
-                    return RawScrollbar(
-                      controller: _scrollController,
-                      thumbVisibility: true,
-                      trackVisibility: true,
-                      thickness: scrollbarThickness,
-                      radius: const Radius.circular(8),
-                      thumbColor: YataColorTokens.textSecondary.withValues(alpha: 0.6),
-                      trackColor: YataColorTokens.divider,
-                      child: GridView.builder(
-                        controller: _scrollController,
-                        primary: false,
-                        padding: const EdgeInsets.only(
-                          right: scrollPad,
-                          bottom: YataSpacingTokens.lg,
-                        ),
-                        gridDelegate: SliverGridDelegateWithFixedCrossAxisCount(
-                          crossAxisCount: safeCrossAxisCount,
-                          crossAxisSpacing: spacing,
-                          mainAxisSpacing: spacing,
-                          mainAxisExtent: tileHeight,
-                        ),
-                        itemCount: items.length,
-                        itemBuilder: (BuildContext context, int index) {
-                          final MenuItemViewData item = items[index];
-                          final bool isSelected = state.isInCart(item.id);
-                          final int? quantity = state.quantityFor(item.id);
-
-                          Widget buildTile() => YataMenuItemTile(
-                            key: ValueKey<String>(item.id),
-                            name: item.name,
-                            priceLabel: state.formatPrice(item.price),
-                            isSelected: isSelected,
-                            minQuantity: 0,
-                            quantity: quantity,
-                            onQuantityChanged: state.isLoading
-                                ? null
-                                : (int value) => controller.updateItemQuantity(item.id, value),
-                            onTap: state.isLoading ? null : () => controller.addMenuItem(item.id),
-                          );
-
-                          if (OrderManagementTracer.shouldSample(index)) {
-                            return OrderManagementTracer.traceSync<Widget>(
-                              "page.menuTile.build",
-                              buildTile,
-                              startArguments: () => <String, dynamic>{
-                                "index": index,
-                                "menuItemId": item.id,
-                                "selected": isSelected,
-                              },
-                              finishArguments: () => <String, dynamic>{"quantity": quantity ?? 0},
-                              logThreshold: const Duration(milliseconds: 2),
-                            );
-                          }
-
-                          return buildTile();
-                        },
-                      ),
-                    );
-                  },
-                  startArguments: () => <String, dynamic>{"maxWidth": constraints.maxWidth},
-                  finishArguments: () => <String, dynamic>{
-                    "itemCount": itemCount,
-                    "crossAxis": sampledCrossAxisCount,
-                    "tileHeight": sampledTileHeight,
-                  },
-                  logThreshold: const Duration(milliseconds: 2),
-                );
-              },
-            ),
-          ),
-        ],
-      ),
-    );
-  }
-}
-
-const double _menuTileMinHeight = 92;
-const double _menuTileTrailingMinHeight = 32 + YataSpacingTokens.md * 2;
-
-double _lineHeight(TextStyle style, TextStyle fallback) {
-  final double baseFontSize = style.fontSize ?? fallback.fontSize ?? 14;
-  final double heightFactor = style.height ?? fallback.height ?? 1.4;
-  return baseFontSize * heightFactor;
-}
-
-double _estimateMenuTileHeight(double itemWidth, BuildContext context) {
-  final TextTheme textTheme = Theme.of(context).textTheme;
-  final TextStyle titleStyle = textTheme.titleMedium ?? YataTypographyTokens.titleMedium;
-  final TextStyle priceStyle = textTheme.bodyMedium ?? YataTypographyTokens.bodyMedium;
-
-  final double titleLineHeight = _lineHeight(titleStyle, YataTypographyTokens.titleMedium);
-  final double priceLineHeight = _lineHeight(priceStyle, YataTypographyTokens.bodyMedium);
-  const double verticalPadding = YataSpacingTokens.md * 2;
-  const double textSpacing = YataSpacingTokens.xs;
-
-  // 横幅が狭い場合は2行になるケースを考慮して高さを確保する。
-  final bool allowTwoLines = itemWidth < 320;
-  final int titleLines = allowTwoLines ? 2 : 1;
-  final double heightFromText =
-      verticalPadding + titleLineHeight * titleLines + textSpacing + priceLineHeight;
-
-  return math.max(math.max(heightFromText, _menuTileTrailingMinHeight), _menuTileMinHeight);
-}
-
-class _CurrentOrderSection extends StatefulWidget {
-  const _CurrentOrderSection({required this.state, required this.controller});
-
-  final OrderManagementState state;
-  final OrderManagementController controller;
-
-  @override
-  State<_CurrentOrderSection> createState() => _CurrentOrderSectionState();
-}
-
-class _CurrentOrderSectionState extends State<_CurrentOrderSection> {
-  final ScrollController _scrollController = ScrollController();
-  final Map<String, GlobalKey> _itemKeys = <String, GlobalKey>{};
-  late final TextEditingController _notesController;
-
-  @override
-  void initState() {
-    super.initState();
-    _notesController = TextEditingController(text: widget.state.orderNotes);
-  }
-
-  @override
-  void didUpdateWidget(covariant _CurrentOrderSection oldWidget) {
-    super.didUpdateWidget(oldWidget);
-    // 自動スクロールは無効化（ハイライトのみ有効）
-    if (widget.state.orderNotes != _notesController.text) {
-      final TextEditingValue oldValue = _notesController.value;
-      final String newText = widget.state.orderNotes;
-      final int maxOffset = newText.length;
-      final int baseOffset = oldValue.selection.baseOffset;
-      final int extentOffset = oldValue.selection.extentOffset;
-      final TextSelection newSelection = TextSelection(
-        baseOffset: baseOffset.clamp(0, maxOffset),
-        extentOffset: extentOffset.clamp(0, maxOffset),
-      );
-      _notesController.value = TextEditingValue(text: newText, selection: newSelection);
-    }
-  }
-
-  @override
-  void dispose() {
-    _scrollController.dispose();
-    _notesController.dispose();
-    super.dispose();
-  }
-
-  // 自動スクロールは無効化
-
-  String _checkoutFailureMessage(CheckoutActionResult result) {
-    switch (result.status) {
-      case CheckoutActionStatus.stockInsufficient:
-        return result.message ?? "在庫が不足している商品があります。数量を調整して再度お試しください。";
-      case CheckoutActionStatus.emptyCart:
-        return result.message ?? "カートに商品がありません。";
-      case CheckoutActionStatus.authenticationFailed:
-        return result.message ?? "ユーザー情報を取得できませんでした。再度ログインしてください。";
-      case CheckoutActionStatus.missingCart:
-        return result.message ?? "カート情報の取得に失敗しました。再度読み込みを行ってください。";
-      case CheckoutActionStatus.failure:
-        return result.message ?? "会計処理に失敗しました。時間をおいて再度お試しください。";
-      case CheckoutActionStatus.success:
-        return "";
-    }
-  }
-
-  @override
-  Widget build(BuildContext context) {
-    final OrderManagementState state = widget.state;
-    final OrderManagementController controller = widget.controller;
-    final TextTheme textTheme = Theme.of(context).textTheme;
-
-    return SizedBox.expand(
-      child: YataSectionCard(
-        expandChild: true,
-        title: "現在の注文",
-        actions: <Widget>[_OrderNumberBadge(orderNumber: state.orderNumber)],
-        child: Column(
-          crossAxisAlignment: CrossAxisAlignment.start,
-          children: <Widget>[
-            // 広幅時のみ列ヘッダを表示
-            LayoutBuilder(
-              builder: (BuildContext context, BoxConstraints constraints) {
-                final bool isNarrow = constraints.maxWidth < 640;
-                if (isNarrow || state.cartItems.isEmpty) {
-                  return const SizedBox.shrink();
-                }
-                final TextStyle headerStyle =
-                    (textTheme.labelMedium ?? YataTypographyTokens.labelMedium).copyWith(
-                      color: YataColorTokens.textSecondary,
-                    );
-                return Column(
-                  crossAxisAlignment: CrossAxisAlignment.stretch,
-                  children: <Widget>[
-                    Row(
-                      children: <Widget>[
-                        Expanded(flex: 6, child: Text("品名", style: headerStyle)),
-                        const SizedBox(width: YataSpacingTokens.md),
-                        Expanded(
-                          flex: 2,
-                          child: Align(
-                            alignment: Alignment.centerRight,
-                            child: Text("単価", style: headerStyle),
-                          ),
-                        ),
-                        const SizedBox(width: YataSpacingTokens.md),
-                        Text("数量", style: headerStyle),
-                        const SizedBox(width: YataSpacingTokens.md),
-                        Expanded(
-                          flex: 2,
-                          child: Align(
-                            alignment: Alignment.centerRight,
-                            child: Text("小計", style: headerStyle),
-                          ),
-                        ),
-                        const SizedBox(width: YataSpacingTokens.sm),
-                        const SizedBox(width: 24),
-                      ],
-                    ),
-                    const SizedBox(height: YataSpacingTokens.xs),
-                    const Divider(
-                      height: YataSpacingTokens.md,
-                      thickness: 1,
-                      color: YataColorTokens.divider,
-                    ),
-                  ],
-                );
-              },
-            ),
-            // リスト領域（内部スクロール）
-            Expanded(
-              child: state.cartItems.isEmpty
-                  ? Center(
-                      child: Padding(
-                        padding: const EdgeInsets.symmetric(vertical: YataSpacingTokens.lg),
-                        child: Text(
-                          "カートに商品が追加されていません",
-                          style: textTheme.bodyMedium?.copyWith(
-                            color: YataColorTokens.textSecondary,
-                          ),
-                        ),
-                      ),
-                    )
-                  : Builder(
-                      builder: (BuildContext context) {
-                        const double scrollbarThickness = 8;
-                        return RawScrollbar(
-                          controller: _scrollController,
-                          thumbVisibility: true,
-                          trackVisibility: true,
-                          thickness: scrollbarThickness,
-                          radius: const Radius.circular(8),
-                          thumbColor: YataColorTokens.textSecondary.withValues(alpha: 0.6),
-                          trackColor: YataColorTokens.divider,
-                          child: ListView.separated(
-                            controller: _scrollController,
-                            padding: const EdgeInsets.only(
-                              right: scrollbarThickness + YataSpacingTokens.xs,
-                            ),
-                            itemCount: state.cartItems.length,
-                            separatorBuilder: (BuildContext context, int index) => const Divider(
-                              height: YataSpacingTokens.lg,
-                              thickness: 1,
-                              color: YataColorTokens.divider,
-                            ),
-                            itemBuilder: (BuildContext context, int index) {
-                              final CartItemViewData item = state.cartItems[index];
-                              _itemKeys.putIfAbsent(item.menuItem.id, GlobalKey.new);
-                              final bool isHighlighted =
-                                  state.highlightedItemId == item.menuItem.id;
-                              return KeyedSubtree(
-                                key: _itemKeys[item.menuItem.id],
-                                child: _HighlightWrapper(
-                                  highlighted: isHighlighted,
-                                  child: _OrderRow(
-                                    name: item.menuItem.name,
-                                    unitPriceLabel: state.formatPrice(item.menuItem.price),
-                                    quantity: item.quantity,
-                                    lineSubtotalLabel: state.formatPrice(item.subtotal),
-                                    onQuantityChanged: (int value) =>
-                                        controller.updateItemQuantity(item.menuItem.id, value),
-                                    onRemove: () => controller.removeItem(item.menuItem.id),
-                                  ),
-                                ),
-                              );
-                            },
-                          ),
-                        );
-                      },
-                    ),
-            ),
-            const SizedBox(height: YataSpacingTokens.md),
-            const Divider(),
-            const SizedBox(height: YataSpacingTokens.md),
-            // 支払い方法セレクター
-            OrderPaymentMethodSelector(
-              selected: state.currentPaymentMethod,
-              isDisabled: state.isCheckoutInProgress || state.isLoading,
-              onChanged: controller.updatePaymentMethod,
-            ),
-            const SizedBox(height: YataSpacingTokens.lg),
-            TextField(
-              controller: _notesController,
-              onChanged: controller.updateOrderNotes,
-              maxLength: 200,
-              enabled: !state.isCheckoutInProgress && !state.isLoading,
-              decoration: InputDecoration(
-                labelText: "メモ",
-                hintText: "例: アレルギー対応、調理指示など",
-                border: OutlineInputBorder(
-                  borderRadius: BorderRadius.circular(YataRadiusTokens.small),
-                  borderSide: const BorderSide(color: YataColorTokens.border),
-                ),
-                focusedBorder: OutlineInputBorder(
-                  borderRadius: BorderRadius.circular(YataRadiusTokens.small),
-                  borderSide: const BorderSide(color: YataColorTokens.primary, width: 2),
-                ),
-                contentPadding: const EdgeInsets.symmetric(
-                  horizontal: YataSpacingTokens.md,
-                  vertical: YataSpacingTokens.sm,
-                ),
-                counterStyle: (textTheme.bodySmall ?? YataTypographyTokens.bodySmall).copyWith(
-                  color: YataColorTokens.textSecondary,
-                ),
-              ),
-              style: textTheme.bodyMedium ?? YataTypographyTokens.bodyMedium,
-            ),
-            const SizedBox(height: YataSpacingTokens.lg),
-            _SummaryRow(
-              label: "小計",
-              value: state.formatPrice(state.subtotal),
-              textTheme: textTheme,
-            ),
-            const SizedBox(height: YataSpacingTokens.xs),
-            _SummaryRow(
-              label: "消費税 (10%)",
-              value: state.formatPrice(state.tax),
-              textTheme: textTheme,
-            ),
-            const SizedBox(height: YataSpacingTokens.md),
-            Row(
-              mainAxisAlignment: MainAxisAlignment.spaceBetween,
-              children: <Widget>[
-                Text("合計", style: textTheme.titleMedium ?? YataTypographyTokens.titleMedium),
-                Text(
-                  state.formatPrice(state.total),
-                  style: textTheme.headlineSmall ?? YataTypographyTokens.headlineSmall,
-                ),
-              ],
-            ),
-            const SizedBox(height: YataSpacingTokens.lg),
-            Row(
-              children: <Widget>[
-                Expanded(
-                  child: OutlinedButton.icon(
-                    onPressed: state.cartItems.isEmpty || state.isCheckoutInProgress
-                        ? null
-                        : controller.clearCart,
-                    icon: const Icon(Icons.close),
-                    label: const Text("クリア"),
-                    style: OutlinedButton.styleFrom(
-                      foregroundColor: YataColorTokens.textPrimary,
-                      side: const BorderSide(color: YataColorTokens.border),
-                      backgroundColor: YataColorTokens.neutral0,
-                      padding: const EdgeInsets.symmetric(vertical: YataSpacingTokens.sm),
-                    ),
-                  ),
-                ),
-                const SizedBox(width: YataSpacingTokens.sm),
-                Expanded(
-                  child: ElevatedButton.icon(
-                    onPressed: state.cartItems.isEmpty || state.isCheckoutInProgress
-                        ? null
-                        : () async {
-                            // 会計処理
-                            final CheckoutActionResult result = await controller.checkout();
-                            if (!context.mounted) {
-                              return;
-                            }
-                            if (result.isSuccess) {
-                              final String? orderNumber = result.order?.orderNumber;
-                              final String orderNumberLabel =
-                                  (orderNumber == null || orderNumber.isEmpty)
-                                  ? "新規注文"
-                                  : "受付コード $orderNumber";
-                              ScaffoldMessenger.of(context).showSnackBar(
-                                SnackBar(content: Text("会計が完了しました（$orderNumberLabel）。")),
-                              );
-                            } else {
-                              final String message = _checkoutFailureMessage(result);
-                              ScaffoldMessenger.of(
-                                context,
-                              ).showSnackBar(SnackBar(content: Text(message)));
-                            }
-                          },
-                    icon: state.isCheckoutInProgress
-                        ? const SizedBox(
-                            width: 18,
-                            height: 18,
-                            child: CircularProgressIndicator(
-                              strokeWidth: 2.2,
-                              valueColor: AlwaysStoppedAnimation<Color>(YataColorTokens.neutral0),
-                            ),
-                          )
-                        : const Icon(Icons.check_circle_outline),
-                    label: Text(state.isCheckoutInProgress ? "会計中…" : "会計"),
-                    style: ElevatedButton.styleFrom(
-                      backgroundColor: YataColorTokens.success,
-                      foregroundColor: YataColorTokens.neutral0,
-                      padding: const EdgeInsets.symmetric(vertical: YataSpacingTokens.sm),
-                    ),
-                  ),
-                ),
-              ],
-            ),
-          ],
-        ),
-      ),
-    );
-  }
-}
-
-class _OrderNumberBadge extends StatelessWidget {
-  const _OrderNumberBadge({required this.orderNumber});
-
-  final String? orderNumber;
-
-  @override
-  Widget build(BuildContext context) {
-    final String label = (orderNumber == null || orderNumber!.isEmpty) ? "割り当て準備中" : orderNumber!;
-    return Container(
-      padding: const EdgeInsets.symmetric(
-        horizontal: YataSpacingTokens.md,
-        vertical: YataSpacingTokens.xs,
-      ),
-      decoration: BoxDecoration(
-        color: YataColorTokens.primarySoft,
-        borderRadius: const BorderRadius.all(Radius.circular(YataRadiusTokens.medium)),
-        border: Border.all(color: YataColorTokens.primary.withValues(alpha: 0.3)),
-      ),
-      child: Text(
-        "受付コード: $label",
-        style:
-            Theme.of(context).textTheme.labelLarge?.copyWith(color: YataColorTokens.primary) ??
-            YataTypographyTokens.labelLarge.copyWith(color: YataColorTokens.primary),
-      ),
-    );
-  }
-}
-
-class _SummaryRow extends StatelessWidget {
-  const _SummaryRow({required this.label, required this.value, required this.textTheme});
-
-  final String label;
-  final String value;
-  final TextTheme textTheme;
-
-  @override
-  Widget build(BuildContext context) => Row(
-    mainAxisAlignment: MainAxisAlignment.spaceBetween,
-    children: <Widget>[
-      Text(
-        label,
-        style: (textTheme.bodyMedium ?? YataTypographyTokens.bodyMedium).copyWith(
-          color: YataColorTokens.textPrimary,
-        ),
-      ),
-      Text(
-        value,
-        style: (textTheme.bodyMedium ?? YataTypographyTokens.bodyMedium).copyWith(
-          color: YataColorTokens.textPrimary,
-          fontWeight: FontWeight.w600,
-        ),
-      ),
-    ],
-  );
-}
-
-class _HighlightWrapper extends StatelessWidget {
-  const _HighlightWrapper({required this.highlighted, required this.child});
-
-  final bool highlighted;
-  final Widget child;
-
-  @override
-  Widget build(BuildContext context) {
-    final Color bgColor = highlighted ? YataColorTokens.selectionTint : Colors.transparent;
-    final Color borderColor = highlighted
-        ? YataColorTokens.primary.withValues(alpha: 0.22)
-        : Colors.transparent;
-    return AnimatedContainer(
-      duration: const Duration(milliseconds: 420),
-      curve: Curves.easeOut,
-      decoration: BoxDecoration(
-        color: bgColor,
-        borderRadius: const BorderRadius.all(Radius.circular(YataRadiusTokens.medium)),
-        border: Border.all(color: borderColor),
-        boxShadow: highlighted
-            ? <BoxShadow>[
-                BoxShadow(
-                  color: YataColorTokens.primary.withValues(alpha: 0.05),
-                  blurRadius: 10,
-                  spreadRadius: 0.5,
-                  offset: const Offset(0, 2),
-                ),
-              ]
-            : const <BoxShadow>[],
-      ),
-      padding: const EdgeInsets.symmetric(
-        horizontal: YataSpacingTokens.sm,
-        vertical: YataSpacingTokens.xs,
-      ),
-      child: child,
-    );
-  }
-}
-
-class _OrderRow extends StatelessWidget {
-  const _OrderRow({
-    required this.name,
-    required this.unitPriceLabel,
-    required this.quantity,
-    required this.lineSubtotalLabel,
-    required this.onQuantityChanged,
-    this.onRemove,
-  });
-
-  final String name;
-  final String unitPriceLabel;
-  final int quantity;
-  final String lineSubtotalLabel;
-  final ValueChanged<int> onQuantityChanged;
-  final VoidCallback? onRemove;
-
-  @override
-  Widget build(BuildContext context) {
-    final TextTheme textTheme = Theme.of(context).textTheme;
-    final TextStyle nameStyle = (textTheme.titleMedium ?? YataTypographyTokens.titleMedium)
-        .copyWith(color: YataColorTokens.textPrimary);
-    final TextStyle priceStyle = (textTheme.bodyMedium ?? YataTypographyTokens.bodyMedium).copyWith(
-      color: YataColorTokens.textSecondary,
-    );
-    final TextStyle subtotalStyle = (textTheme.bodyMedium ?? YataTypographyTokens.bodyMedium)
-        .copyWith(color: YataColorTokens.textPrimary, fontWeight: FontWeight.w600);
-
-    return Padding(
-      padding: const EdgeInsets.symmetric(vertical: YataSpacingTokens.sm),
-      child: LayoutBuilder(
-        builder: (BuildContext context, BoxConstraints constraints) {
-          final bool isNarrow = constraints.maxWidth < 640;
-
-          if (isNarrow) {
-            // 2行レイアウト: 1行目=品名 | 単価 | ×, 2行目= [ステッパー] | 行小計
-            return Column(
-              crossAxisAlignment: CrossAxisAlignment.start,
-              children: <Widget>[
-                Row(
-                  children: <Widget>[
-                    Expanded(
-                      child: Text(name, style: nameStyle, overflow: TextOverflow.ellipsis),
-                    ),
-                    const SizedBox(width: YataSpacingTokens.sm),
-                    Text(unitPriceLabel, style: priceStyle, overflow: TextOverflow.fade),
-                    const SizedBox(width: YataSpacingTokens.sm),
-                    if (onRemove != null)
-                      IconButton(
-                        onPressed: onRemove,
-                        icon: const Icon(Icons.close),
-                        tooltip: "削除",
-                        splashRadius: 18,
-                        color: YataColorTokens.textSecondary,
-                      ),
-                  ],
-                ),
-                const SizedBox(height: YataSpacingTokens.xs),
-                Row(
-                  children: <Widget>[
-                    YataQuantityStepper(
-                      value: quantity,
-                      onChanged: onQuantityChanged,
-                      compact: true,
-                    ),
-                    const SizedBox(width: YataSpacingTokens.md),
-                    Expanded(
-                      child: _SubtotalDisplay(
-                        value: lineSubtotalLabel,
-                        textStyle: subtotalStyle,
-                        quantity: quantity,
-                      ),
-                    ),
-                  ],
-                ),
-              ],
-            );
-          }
-
-          // 1行レイアウト（できるだけ品名を優先して表示）
-          return Row(
-            children: <Widget>[
-              // 品名
-              Expanded(
-                flex: 6,
-                child: Text(name, style: nameStyle, maxLines: 2, overflow: TextOverflow.ellipsis),
-              ),
-              const SizedBox(width: YataSpacingTokens.md),
-              // 単価
-              Expanded(
-                flex: 2,
-                child: Align(
-                  alignment: Alignment.centerRight,
-                  child: Text(unitPriceLabel, style: priceStyle, overflow: TextOverflow.ellipsis),
-                ),
-              ),
-              const SizedBox(width: YataSpacingTokens.md),
-              // 数量ステッパー（非フレックス）
-              YataQuantityStepper(value: quantity, onChanged: onQuantityChanged, compact: true),
-              const SizedBox(width: YataSpacingTokens.md),
-              // 行小計
-              Expanded(
-                flex: 2,
-                child: _SubtotalDisplay(
-                  value: lineSubtotalLabel,
-                  textStyle: subtotalStyle,
-                  quantity: quantity,
-                ),
-              ),
-              const SizedBox(width: YataSpacingTokens.sm),
-              // 削除ボタン
-              if (onRemove != null)
-                IconButton(
-                  onPressed: onRemove,
-                  icon: const Icon(Icons.close),
-                  tooltip: "削除",
-                  splashRadius: 18,
-                  color: YataColorTokens.textSecondary,
-                ),
-            ],
-          );
-        },
-      ),
-    );
-  }
-}
-
-class _SubtotalDisplay extends StatelessWidget {
-  const _SubtotalDisplay({required this.value, required this.textStyle, required this.quantity});
-
-  final String value;
-  final TextStyle textStyle;
-  final int quantity;
-
-  @override
-  Widget build(BuildContext context) => Align(
-    alignment: Alignment.centerRight,
-    child: Row(
-      mainAxisAlignment: MainAxisAlignment.end,
-      children: <Widget>[
-        Text("x$quantity", style: textStyle),
-        const SizedBox(width: YataSpacingTokens.sm),
-        Icon(Icons.arrow_forward, size: 18, color: YataColorTokens.textSecondary),
-        const SizedBox(width: YataSpacingTokens.xs),
-        Flexible(
-          child: Text(value, style: textStyle, overflow: TextOverflow.ellipsis, softWrap: false),
-        ),
-      ],
-    ),
-  );
-}
-
-class _OrderPageErrorBanner extends StatelessWidget {
-  const _OrderPageErrorBanner({required this.message, required this.onRetry});
-
-  final String message;
-  final VoidCallback onRetry;
-
-  @override
-  Widget build(BuildContext context) => Container(
-    padding: const EdgeInsets.all(YataSpacingTokens.md),
-    decoration: BoxDecoration(
-      color: YataColorTokens.dangerSoft,
-      borderRadius: const BorderRadius.all(Radius.circular(YataRadiusTokens.medium)),
-      border: Border.all(color: YataColorTokens.danger.withValues(alpha: 0.3)),
-    ),
-    child: Row(
-      children: <Widget>[
-        const Icon(Icons.error_outline, color: YataColorTokens.danger),
-        const SizedBox(width: YataSpacingTokens.sm),
-        Expanded(
-          child: Text(
-            message,
-            style: Theme.of(context).textTheme.bodyMedium?.copyWith(
-              color: YataColorTokens.danger,
-              fontWeight: FontWeight.w600,
-            ),
-          ),
-        ),
-        TextButton.icon(
-          onPressed: onRetry,
-          icon: const Icon(Icons.refresh),
-          label: const Text("再試行"),
-        ),
-      ],
-    ),
-  );
 }
